@@ -1,6 +1,6 @@
 {
   "name": "r6-api-caching",
-  "version": "1.4.0",
+  "version": "1.5.0",
   "description": "Rainbow Six Siege data caching",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -20,14 +20,13 @@
   },
   "homepage": "https://github.com/PoloLacoste/r6-api-caching#readme",
   "devDependencies": {
-    "@types/mongodb": "^3.6.20",
-    "@types/node": "^14.17.5",
-    "typescript": "^4.1.3"
+    "@types/node": "^17.0.31",
+    "typescript": "^4.6.4"
   },
   "dependencies": {
-    "class-validator": "^0.13.1",
-    "mongodb": "^4.0.0",
-    "r6api.js": "^4.0.0",
-    "redis": "^3.1.2"
+    "class-validator": "^0.13.2",
+    "mongodb": "^4.5.0",
+    "r6api.js": "^4.1.0",
+    "redis": "^4.1.0"
   }
 }
